// src/api/stetsom/model/bannersPayload.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { Banner } from "./banner";

export interface BannersPayload {
  items: Banner[];
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  total: number;
}