// src/api/stetsom/model/bannerWithUploads.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { Banner } from "./banner";
import type { BannerWithUploadsUploads } from "./bannerWithUploadsUploads";

export interface BannerWithUploads {
  banner: Banner;
  uploads: BannerWithUploadsUploads;
}