// src/api/stetsom/model/uploadPresignResponse.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { UploadPresignResponseAssetType } from "./uploadPresignResponseAssetType";
import type { UploadPresignResponseHeaders } from "./uploadPresignResponseHeaders";
import type { UploadPresignResponseMethod } from "./uploadPresignResponseMethod";

export interface UploadPresignResponse {
  /** PUT your file here — signature is embedded */
  uploadUrl: string;
  /** Permanent public URL — store in resource fields */
  file_url: string;
  key: string;
  method: UploadPresignResponseMethod;
  /**
   * Presigned URL TTL in seconds
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  expiresIn: number;
  /** Include in the PUT request */
  headers: UploadPresignResponseHeaders;
  assetType: UploadPresignResponseAssetType;
  fileName: string;
}